Refactoring code of cards on the Data Contracts page
