+++ packages/frontend/src/components/dataContracts/index.js
@@ -1,9 +1,9 @@
 import TrendingDataContracts from './TrendingDataContracts'
 import DataContractsList from './DataContractsList'
-import TopDataContracts from './TopDataContracts'
+import { CardsBlock } from './Cards'
 
 export {
   TrendingDataContracts,
   DataContractsList,
-  TopDataContracts
+  CardsBlock
 }
